requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
pillow>=10.0.0,<12.0.0
pytesseract>=0.3.10,<1.0.0
